+++ src/renderer/TextureDesc.ts
@@ -1,0 +1,8 @@
+export interface TextureDesc {
+    name: string;
+    format: GPUTextureFormat;
+    width: number;
+    height: number;
+    usage: GPUTextureUsageFlags;
+    sampleCount?: number;
+}
